+++ src/AbyssIrc.Core/Events/Core/ServerStoppingEvent.cs
@@ -1,0 +1,3 @@
+namespace AbyssIrc.Core.Events.Core;
+
+public record ServerStoppingEvent;