+++ src/AbyssIrc.Server/Listeners/Base/BaseHandler.cs
@@ -1,10 +1,14 @@
+using AbyssIrc.Core.Data.Configs;
 using AbyssIrc.Network.Interfaces.Commands;
 using AbyssIrc.Server.Data.Events;
 using AbyssIrc.Server.Data.Events.Irc;
 using AbyssIrc.Server.Data.Events.TcpServer;
+using AbyssIrc.Server.Data.Internal.Handlers;
 using AbyssIrc.Server.Data.Internal.Sessions;
 using AbyssIrc.Server.Interfaces.Services.System;
+using AbyssIrc.Signals.Interfaces.Listeners;
 using AbyssIrc.Signals.Interfaces.Services;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Serilog;
 using ILogger = Microsoft.Extensions.Logging.ILogger;
@@ -15,22 +19,37 @@ public abstract class BaseHandler
 {
     protected ILogger Logger { get; }
 
+    protected AbyssServerData ServerData { get; }
+
+    protected AbyssIrcConfig ServerConfig { get; }
+
     private readonly IAbyssSignalService _signalService;
 
     private readonly ISessionManagerService _sessionManagerService;
 
+    protected string Hostname => ServerData.Hostname;
+
     protected BaseHandler(
-        ILogger<BaseHandler> logger, IAbyssSignalService signalService, ISessionManagerService sessionManagerService
+        ILogger<BaseHandler> logger, IServiceProvider serviceProvider
     )
     {
         Logger = logger;
-        _signalService = signalService;
-        _sessionManagerService = sessionManagerService;
+        _signalService = serviceProvider.GetRequiredService<IAbyssSignalService>();
+        _sessionManagerService = serviceProvider.GetRequiredService<ISessionManagerService>();
+        ServerData = serviceProvider.GetRequiredService<AbyssServerData>();
+        ServerConfig = serviceProvider.GetRequiredService<AbyssIrcConfig>();
     }
 
     protected Task SendIrcMessageAsync(string id, IIrcCommand message)
     {
         return _signalService.PublishAsync(new SendIrcMessageEvent(id, message));
+    }
+
+
+    protected void SubscribeSignal<TEvent>(IAbyssSignalListener<TEvent> listener)
+        where TEvent : class
+    {
+        _signalService.Subscribe(listener);
     }
 
 
@@ -40,12 +59,17 @@ public abstract class BaseHandler
     }
 
 
+    protected List<IrcSession> GetSessions()
+    {
+        return _sessionManagerService.GetSessions();
+    }
+
     /// <summary>
     ///   Get the session by the id
     /// </summary>
     /// <param name="id"></param>
     /// <returns></returns>
-    protected IrcSession GetSession(string id)
+    protected IrcSession? GetSession(string id)
     {
         return _sessionManagerService.GetSession(id);
     }